SLDB songlength, startTime, fadeIn, fadeOut is now millisecond exact
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2233 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/builder/netsiddev/NetSIDDevBuilder.java
@@ -105,13 +105,13 @@ public class NetSIDDevBuilder implements SIDBuilder, Mixer {
 	}
 
 	@Override
-	public void fadeIn(int fadeIn) {
+	public void fadeIn(double fadeIn) {
 		System.err.println("Fade-in unsupported by network SID client");
 		// XXX unsupported by JSIDDevice
 	}
 
 	@Override
-	public void fadeOut(int fadeOut) {
+	public void fadeOut(double fadeOut) {
 		System.err.println("Fade-out unsupported by network SID client");
 		// XXX unsupported by JSIDDevice
 	}
